Compile fix

+++ include/Blob/GL/VertexBufferObject.hpp
@@ -12,12 +12,12 @@ private:
     uint32_t vertexBufferObject = 0;
 
 protected:
-    int64_t dataSize = 0;
+    size_t dataSize = 0;
 
 public:
     VertexBufferObject();
 
-    VertexBufferObject(const uint8_t *data, int64_t dataSize);
+    VertexBufferObject(const uint8_t *data, size_t dataSize);
 
     VertexBufferObject(const VertexBufferObject &) = delete;
 
@@ -25,8 +25,8 @@ public:
 
     ~VertexBufferObject();
 
-    void setData(const uint8_t *data, int64_t dataSize, bool dynamic = false);
+    void setData(const uint8_t *data, size_t dataSize, bool dynamic = false);
 
-    void setSubData(uint8_t *data, int64_t dataSize, int64_t offset = 0);
+    void setSubData(uint8_t *data, size_t dataSize, size_t offset = 0);
 };
 } // namespace Blob::GL
